+++ src/main/java/com/curso/angular/service/interfaces/IClienteService.java
@@ -3,6 +3,6 @@ package com.curso.angular.service.interfaces;
 import com.curso.angular.base.service.interfaces.IServiceCrudAbstract;
 import com.curso.angular.entity.Cliente;
 
-public interface IClienteService extends IServiceCrudAbstract<Cliente, Long>{
+public interface IClienteService extends IServiceCrudAbstract<Cliente, Integer>{
 
 }
